Extract challenge into shared BaseAPI
