+++ module/src/main/java/com/authsignal/push/AuthsignalPush.kt
@@ -59,7 +59,7 @@ class AuthsignalPush(
     var deviceIntegrityToken: String? = null
 
     if (performAttestation) {
-      val challengeResponse = api.challenge(userToken)
+      val challengeResponse = api.challenge(token = userToken)
 
       val nonce = challengeResponse.data?.nonce
         ?: return AuthsignalResponse(
