+++ module/src/main/java/com/authsignal/qr/api/QRCodeAPI.kt
@@ -1,8 +1,8 @@
 package com.authsignal.qr.api
 
 import com.authsignal.APIError
+import com.authsignal.BaseAPI
 import com.authsignal.Encoder
-import com.authsignal.HttpClientFactory
 import com.authsignal.models.AuthsignalResponse
 import com.authsignal.models.api.*
 import com.authsignal.models.*
@@ -12,10 +12,7 @@ import io.ktor.client.call.*
 import io.ktor.client.request.*
 import io.ktor.http.*
 
-class QRCodeAPI(tenantID: String, private val baseURL: String) {
-  private val client = HttpClientFactory.create()
-
-  private val basicAuth = "Basic ${Encoder.toBase64String("$tenantID:".toByteArray())}"
+class QRCodeAPI(tenantID: String, baseURL: String) : BaseAPI(tenantID, baseURL) {
 
   suspend fun getCredential(publicKey: String): AuthsignalResponse<AppCredential> {
     val encodedKey = Encoder.toBase64String(publicKey.toByteArray())
@@ -112,32 +109,6 @@ class QRCodeAPI(tenantID: String, private val baseURL: String) {
 
       return if (success) {
         AuthsignalResponse(data = true)
-      } else {
-        APIError.mapToErrorResponse(response)
-      }
-    } catch (e: Exception) {
-      APIError.handleNetworkException(e)
-    }
-  }
-
-  suspend fun challenge(token: String): AuthsignalResponse<ChallengeResponse> {
-    val url = "$baseURL/client/challenge"
-    val body = ChallengeRequest(action = null)
-
-    return try {
-      val response = client.post(url) {
-        contentType(ContentType.Application.Json)
-        setBody(body)
-
-        headers {
-          append(HttpHeaders.Authorization, "Bearer $token")
-        }
-      }
-
-      if (response.status == HttpStatusCode.OK) {
-        val data = response.body<ChallengeResponse>()
-
-        AuthsignalResponse(data = data)
       } else {
         APIError.mapToErrorResponse(response)
       }
